refactor(agent): extract JsonRpcDispatcher + AgentToolExecutor; AgentServer becomes a thin facade
The protocol layer (initialize/ping/tools/list/tools/call routing, JSON-RPC
response shapes) moves to an instance JsonRpcDispatcher; the tool-execution
layer (emergency-stop/AgentCommandsEnabled/per-command gates, argument
validation, ToolCatalog command building, the #113/#105 dispatch rules, the
meta-tools, and #101 envelope publication) moves to an instance
AgentToolExecutor taking settings/invoker/session accessors via constructor.

AgentServer shrinks to a static facade that wires the production instances
against AgentRuntime and re-exposes the members hosts and tests compile
against (Dispatch, RunStdio, StartHttp/StopHttp/IsHttpListening,
Instructions, SerializesOnInputLock, the canonical constants). Behavior is
unchanged; no test seams remain beyond InternalsVisibleTo forwarding.

Part of #215.
